👌 IMPROVE: Refactorization et suppression de MUI
