+++ package.json
@@ -19,14 +19,9 @@
     "package:dist": "electron-builder --dir"
   },
   "keywords": [],
-  "author": {
-    "name": "Sebastien Colbert"
-  },
+  "author": "HelloSoftware",
   "license": "MIT",
   "dependencies": {
-    "@emotion/react": "^11.9.0",
-    "@emotion/styled": "^11.8.1",
-    "@mui/material": "^5.6.4",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "uuid": "^8.3.2"
