+++ src/renderer/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import Slider from './Slider';
+import Slider from './components/Slider';
 
 export default function App(): JSX.Element {
   return (
